rgw/mdlog: move data types to rgw_mdlog_types.h
consolidate the generic mdlog types and their encodings
