+++ src/rgw/CMakeLists.txt
@@ -67,6 +67,7 @@ set(librgw_common_srcs
   rgw_lc.cc
   rgw_lc_s3.cc
   rgw_restore.cc
+  rgw_mdlog_types.cc
   rgw_metadata.cc
   rgw_multi.cc
   rgw_multi_del.cc
