+++ src/rgw/rgw_metadata.cc
@@ -9,106 +9,6 @@
 
 using namespace std;
 
-void LogStatusDump::dump(Formatter *f) const {
-  string s;
-  switch (status) {
-    case MDLOG_STATUS_WRITE:
-      s = "write";
-      break;
-    case MDLOG_STATUS_SETATTRS:
-      s = "set_attrs";
-      break;
-    case MDLOG_STATUS_REMOVE:
-      s = "remove";
-      break;
-    case MDLOG_STATUS_COMPLETE:
-      s = "complete";
-      break;
-    case MDLOG_STATUS_ABORT:
-      s = "abort";
-      break;
-    default:
-      s = "unknown";
-      break;
-  }
-  encode_json("status", s, f);
-}
-
-void encode_json(const char *name, const obj_version& v, Formatter *f)
-{
-  f->open_object_section(name);
-  f->dump_string("tag", v.tag);
-  f->dump_unsigned("ver", v.ver);
-  f->close_section();
-}
-
-void decode_json_obj(obj_version& v, JSONObj *obj)
-{
-  JSONDecoder::decode_json("tag", v.tag, obj);
-  JSONDecoder::decode_json("ver", v.ver, obj);
-}
-
-void RGWMetadataLogData::encode(bufferlist& bl) const {
-  ENCODE_START(1, 1, bl);
-  encode(read_version, bl);
-  encode(write_version, bl);
-  uint32_t s = (uint32_t)status;
-  encode(s, bl);
-  ENCODE_FINISH(bl);
-}
-
-void RGWMetadataLogData::decode(bufferlist::const_iterator& bl) {
-   DECODE_START(1, bl);
-   decode(read_version, bl);
-   decode(write_version, bl);
-   uint32_t s;
-   decode(s, bl);
-   status = (RGWMDLogStatus)s;
-   DECODE_FINISH(bl);
-}
-
-void RGWMetadataLogData::dump(Formatter *f) const {
-  encode_json("read_version", read_version, f);
-  encode_json("write_version", write_version, f);
-  encode_json("status", LogStatusDump(status), f);
-}
-
-void decode_json_obj(RGWMDLogStatus& status, JSONObj *obj) {
-  string s;
-  JSONDecoder::decode_json("status", s, obj);
-  if (s == "complete") {
-    status = MDLOG_STATUS_COMPLETE;
-  } else if (s == "write") {
-    status = MDLOG_STATUS_WRITE;
-  } else if (s == "remove") {
-    status = MDLOG_STATUS_REMOVE;
-  } else if (s == "set_attrs") {
-    status = MDLOG_STATUS_SETATTRS;
-  } else if (s == "abort") {
-    status = MDLOG_STATUS_ABORT;
-  } else {
-    status = MDLOG_STATUS_UNKNOWN;
-  }
-}
-
-void RGWMetadataLogData::decode_json(JSONObj *obj) {
-  JSONDecoder::decode_json("read_version", read_version, obj);
-  JSONDecoder::decode_json("write_version", write_version, obj);
-  JSONDecoder::decode_json("status", status, obj);
-}
-
-std::list<RGWMetadataLogData> RGWMetadataLogData::generate_test_instances() {
-  std::list<RGWMetadataLogData> l;
-  l.emplace_back();
-  l.emplace_back();
-  l.back().read_version = obj_version();
-  l.back().read_version.tag = "read_tag";
-  l.back().write_version = obj_version();
-  l.back().write_version.tag = "write_tag";
-  l.back().status = MDLOG_STATUS_WRITE;
-  return l;
-}
-
 class RGWMetadataTopHandler : public RGWMetadataHandler {
   struct iter_data {
     set<string> sections;
